# app/svc/docs/swagger.yaml
basePath: /
definitions:
...
  handler.Annotation:
    properties:
      authorUserId:
        type: string
      createdAt:
        type: integer
      entryId:
        type: string
      id:
        type: string
      sharedWith:
        type: string
      text:
        type: string
      updatedAt:
        type: integer
      visibility:
        type: string
    type: object
  handler.AuthForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  handler.AuthResetPasswordRequest:
    properties:
      newPassword:
        type: string
      token:
        type: string
    type: object
...
  handler.CreateAnnotationRequest:
    properties:
      entryId:
        type: string
      sharedWith:
        type: string
      text:
        type: string
      visibility:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.OkResponse:
    properties:
      ok:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 'Backend-for-frontend: session management, campaigns, entries, annotations.'
  title: Quillit BFF API
  version: "1.0"
paths:
...
  /api/annotations:
    get:
      parameters:
      - description: Filter by entry ID
        in: query
        name: entryId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Annotation'
            type: array
      summary: List annotations
      tags:
      - annotations
    post:
      consumes:
      - application/json
      parameters:
      - description: Annotation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAnnotationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Annotation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create annotation
      tags:
      - annotations
  /api/annotations/{id}:
    delete:
      parameters:
      - description: Annotation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Delete annotation
      tags:
      - annotations
    patch:
      consumes:
      - application/json
      description: Partial update — send only the fields to change.
      parameters:
      - description: Annotation ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAnnotationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Annotation'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update annotation
      tags:
      - annotations
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Proxies to auth-svc. Always returns 200 regardless of whether the
        email is registered.
      parameters:
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Forgot password
      tags:
      - auth
...
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Proxies to auth-svc.
      parameters:
      - description: Token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Reset password
      tags:
      - auth
...
swagger: "2.0"